+++ lib/caylir/graph/request.ex
@@ -3,11 +3,67 @@ defmodule Caylir.Graph.Request do
   Sends requests to the server.
   """
 
+  alias Caylir.Graph
+
   @doc """
-  Send a request to the server.
+  Deletes a quad from the graph.
   """
-  @spec send({ atom, String.t, String.t }) :: any
-  def send({ method, url, payload }) do
+  @spec delete(Keyword.t, Keyword.t) :: Graph.t_delete
+  def delete(quad, conn) do
+    url  = Graph.URL.delete(conn)
+    body = [ quad ] |> Poison.encode!
+
+    case send(:post, url, body) do
+      { :ok, 200, _success }          -> :ok
+      { :ok, 400, %{ error: reason }} -> { :error, reason }
+    end
+  end
+
+  @doc """
+  Queries the graph.
+  """
+  @spec query(String.t, Keyword.t) :: Graph.t_query
+  def query(query, conn) do
+    url = Graph.URL.query(conn)
+
+    case send(:post, url, query) do
+      { :ok, 200, %{ result: result }} -> result
+      { :ok, 400, %{ error:  reason }} -> { :error, reason }
+    end
+  end
+
+
+  @doc """
+  Gets the shape of a query.
+  """
+  @spec shape(String.t, Keyword.t) :: Graph.t_query
+  def shape(query, conn) do
+    url = Graph.URL.shape(conn)
+
+    case send(:post, url, query) do
+      { :ok, 200, shape }             -> shape
+      { :ok, 400, %{ error: reason }} -> { :error, reason }
+    end
+  end
+
+  @doc """
+  Writes a quad to the graph.
+  """
+  @spec write(Keyword.t, Keyword.t) :: Graph.t_write
+  def write(quad, conn) do
+    url  = Graph.URL.write(conn)
+    body = [ quad ] |> Poison.encode!
+
+    case send(:post, url, body) do
+      { :ok, 200, _success }          -> :ok
+      { :ok, 400, %{ error: reason }} -> { :error, reason }
+    end
+  end
+
+
+  # Utility methods
+
+  defp send(method, url, payload) do
     body    = payload |> :binary.bin_to_list()
     headers = [{ 'Content-Type',  'application/x-www-form-urlencoded' },
                { 'Content-Length', length(body) }]
@@ -17,7 +73,6 @@ defmodule Caylir.Graph.Request do
 
     parse_response(status, response)
   end
-
 
   defp parse_response(status, ''),  do: { :ok, status }
   defp parse_response(status, body) do